address review comments

+++ emily/handler/tests/integration/withdrawal.rs
@@ -143,7 +143,7 @@ async fn get_withdrawals() {
     // If the number of elements is an exact multiple of the chunk size the "final"
     // query will still have a next token, and the next query will now have a next
     // token and will return no additional data.
-    let expected_chunks = expected_withdrawal_infos.len() as u32 / chunksize + 1;
+    let expected_chunks = expected_withdrawal_infos.len() / chunksize + 1;
 
     // Act.
     // ----
@@ -157,7 +157,7 @@ async fn get_withdrawals() {
             &configuration,
             status,
             next_token.as_ref().and_then(|o| o.as_deref()),
-            Some(chunksize),
+            Some(chunksize as u32),
         )
         .await
         .expect("Received an error after making a valid get withdrawal api call.");
@@ -171,13 +171,13 @@ async fn get_withdrawals() {
 
     // Assert.
     // -------
-    assert_eq!(expected_chunks, gotten_withdrawal_info_chunks.len() as u32);
+    assert_eq!(expected_chunks, gotten_withdrawal_info_chunks.len());
     let max_chunk_size = gotten_withdrawal_info_chunks
         .iter()
         .map(|chunk| chunk.len())
         .max()
         .unwrap();
-    assert!(chunksize >= max_chunk_size as u32);
+    assert!(chunksize >= max_chunk_size);
 
     let mut gotten_withdrawal_infos = gotten_withdrawal_info_chunks
         .into_iter()
